logic for recovering the password
