+++ src/main/java/com/wuav/client/gui/controllers/AllUsersController.java
@@ -124,7 +124,6 @@ public class AllUsersController  extends RootController implements Initializable
             stage.getProperties().put("previousStage", previousStage);
         });
 
-
         stage.setResizable(false);
         stage.setScene(scene);
         stage.show();
