+++ src/main/java/com/wuav/client/gui/controllers/LoginController.java
@@ -50,72 +50,22 @@ public class LoginController extends RootController implements Initializable {
     private MFXButton login;
 
     @FXML
-    private MFXButton recoverPassword;
-    @FXML
     private StackPane baseContent;
     private final IControllerFactory controllerFactory;
 
     private final IAuthService authService;
-    private final IEmailEngine emailEngine;
-    private final IEmailSender emailSender;
-
-
-
-
 
     @Inject
-    public LoginController(IControllerFactory controllerFactory, IAuthService authService, IEmailEngine emailEngine, IEmailSender emailSender) {
+    public LoginController(IControllerFactory controllerFactory, IAuthService authService) {
         this.controllerFactory = controllerFactory;
         this.authService = authService;
-        this.emailEngine = emailEngine;
-        this.emailSender = emailSender;
+
     }
 
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        recoverPassword.setOnAction(e -> recoverPasswordByEmail());
     }
-
-    private void recoverPasswordByEmail() {
-        Session session = EmailConnectionFactory.getSession();
-
-        String newPassword = authService.generateNewPassword(userEmailField.getText());
-
-        if (newPassword.equals("WrongEmail")) {
-          AlertHelper.showDefaultAlert("User with email " + userEmailField.getText() + "was not found", Alert.AlertType.ERROR);
-
-        } else {
-        new Thread(() -> {
-            try {
-
-                // Define the template name and variables
-                String templateName = "email-template-confirm";
-                Map<String, Object> templateVariables = new HashMap<>();
-                templateVariables.put("newPassword", newPassword);
-
-                // Process the template and generate the email body
-                String emailBody = emailEngine.processTemplate(templateName, templateVariables);
-                // Update the UI on the JavaFX application thread
-                var emailResult = emailSender.sendEmail(session, "vince.kautzer@ethereal.email","Installation completed", emailBody,false,null);
-                Platform.runLater(() -> {
-                    // Display message
-                    if (emailResult) {
-                        AlertHelper.showDefaultAlert("Email successfully sent with new password ", Alert.AlertType.INFORMATION);
-
-                    }
-                });
-            } catch (Exception e) {
-                // Handle sending failure
-                Platform.runLater(() -> {
-                    // Show an error message
-                    AlertHelper.showDefaultAlert("Email sending failed " + e.getMessage(), Alert.AlertType.ERROR);
-                });
-            }
-        }).start();
-        }
-    }
-
 
     private RootController loadNodesView(ViewType viewType) throws IOException {
         return controllerFactory.loadFxmlFile(viewType);
